convert new_phone_number_detection to phone_number_detection
